+++ shiki_organizer/gui/ui/main_window.py
@@ -15,9 +15,8 @@ from PySide6.QtGui import (QBrush, QColor, QConicalGradient, QCursor,
     QFont, QFontDatabase, QGradient, QIcon,
     QImage, QKeySequence, QLinearGradient, QPainter,
     QPalette, QPixmap, QRadialGradient, QTransform)
-from PySide6.QtWidgets import (QAbstractItemView, QApplication, QHBoxLayout, QHeaderView,
-    QMainWindow, QPushButton, QSizePolicy, QSpacerItem,
-    QTableView, QVBoxLayout, QWidget)
+from PySide6.QtWidgets import (QApplication, QMainWindow, QSizePolicy, QTabWidget,
+    QVBoxLayout, QWidget)
 
 class Ui_MainWindow(object):
     def setupUi(self, MainWindow):
@@ -28,69 +27,22 @@ class Ui_MainWindow(object):
         self.centralwidget.setObjectName(u"centralwidget")
         self.verticalLayout = QVBoxLayout(self.centralwidget)
         self.verticalLayout.setObjectName(u"verticalLayout")
-        self.view = QTableView(self.centralwidget)
-        self.view.setObjectName(u"view")
-        self.view.setSelectionBehavior(QAbstractItemView.SelectRows)
+        self.tab_widget = QTabWidget(self.centralwidget)
+        self.tab_widget.setObjectName(u"tab_widget")
 
-        self.verticalLayout.addWidget(self.view)
-
-        self.horizontalLayout_3 = QHBoxLayout()
-        self.horizontalLayout_3.setObjectName(u"horizontalLayout_3")
-        self.horizontalLayout_2 = QHBoxLayout()
-        self.horizontalLayout_2.setObjectName(u"horizontalLayout_2")
-        self.pushButton_4 = QPushButton(self.centralwidget)
-        self.pushButton_4.setObjectName(u"pushButton_4")
-
-        self.horizontalLayout_2.addWidget(self.pushButton_4)
-
-        self.pushButton_5 = QPushButton(self.centralwidget)
-        self.pushButton_5.setObjectName(u"pushButton_5")
-
-        self.horizontalLayout_2.addWidget(self.pushButton_5)
-
-
-        self.horizontalLayout_3.addLayout(self.horizontalLayout_2)
-
-        self.horizontalSpacer = QSpacerItem(40, 20, QSizePolicy.Expanding, QSizePolicy.Minimum)
-
-        self.horizontalLayout_3.addItem(self.horizontalSpacer)
-
-        self.horizontalLayout = QHBoxLayout()
-        self.horizontalLayout.setObjectName(u"horizontalLayout")
-        self.button_start = QPushButton(self.centralwidget)
-        self.button_start.setObjectName(u"button_start")
-
-        self.horizontalLayout.addWidget(self.button_start)
-
-        self.button_stop = QPushButton(self.centralwidget)
-        self.button_stop.setObjectName(u"button_stop")
-
-        self.horizontalLayout.addWidget(self.button_stop)
-
-        self.button_done = QPushButton(self.centralwidget)
-        self.button_done.setObjectName(u"button_done")
-
-        self.horizontalLayout.addWidget(self.button_done)
-
-
-        self.horizontalLayout_3.addLayout(self.horizontalLayout)
-
-
-        self.verticalLayout.addLayout(self.horizontalLayout_3)
+        self.verticalLayout.addWidget(self.tab_widget)
 
         MainWindow.setCentralWidget(self.centralwidget)
 
         self.retranslateUi(MainWindow)
+
+        self.tab_widget.setCurrentIndex(1)
+
 
         QMetaObject.connectSlotsByName(MainWindow)
     # setupUi
 
     def retranslateUi(self, MainWindow):
         MainWindow.setWindowTitle(QCoreApplication.translate("MainWindow", u"Shiki Organizer", None))
-        self.pushButton_4.setText(QCoreApplication.translate("MainWindow", u"Add (A)", None))
-        self.pushButton_5.setText(QCoreApplication.translate("MainWindow", u"Delete (D)", None))
-        self.button_start.setText(QCoreApplication.translate("MainWindow", u"Start (S)", None))
-        self.button_stop.setText(QCoreApplication.translate("MainWindow", u"Stop (Q)", None))
-        self.button_done.setText(QCoreApplication.translate("MainWindow", u"Complete (C)", None))
     # retranslateUi