+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 
 - so-repository modify command.
 - so-repository pull command show info about its own work.
-- basic gui.
+- basic GUI with "Today" and "Tasks" tabs ([#13](https://github.com/nikohonu/shiki-organizer/issues/13), [#18](https://github.com/nikohonu/shiki-organizer/issues/18)).
 - start, stop, done commands for gui.
 - show new scheduled date when user complete recurring tasks.
 
@@ -19,7 +19,7 @@
 ### Fixed
 
 - so-repository ls command.
-- so-interval command [#17](https://github.com/nikohonu/shiki-organizer/issues/17).
+- so-interval command ([#17](https://github.com/nikohonu/shiki-organizer/issues/17)).
 
 ## [0.6.0] - 2022-12-13